requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
psycopg2-binary  
networkx
python-dotenv
